PYIC-2867: Move Verifiable Credential Classes to Library
