+++ libs/verifiable-credentials/src/main/java/uk/gov/di/ipv/core/library/verifiablecredential/exception/VerifiableCredentialException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.retrievecricredential.exception;
+package uk.gov.di.ipv.core.library.verifiablecredential.exception;
 
 import uk.gov.di.ipv.core.library.domain.ErrorResponse;
