+++ settings.gradle
@@ -20,4 +20,5 @@ include "lib",
 		"integration-test",
 		"libs:credential-issuer-config-service",
 		"libs:kms-es256-signer",
-		"libs:vc-helper"
+		"libs:vc-helper",
+		"libs:verifiable-credentials"
